requirements.txt
fastapi
prisma
uvicorn
aiohttp
pillow